refactor(frontend): use entities index import
