+++ frontend/app/.server/domain/mappers/client-friendly-status.dto.mapper.ts
@@ -1,7 +1,7 @@
 import { injectable } from 'inversify';
 
 import type { ClientFriendlyStatusDto } from '~/.server/domain/dtos';
-import type { ClientFriendlyStatusEntity } from '~/.server/domain/entities/client-friendly-status.entity';
+import type { ClientFriendlyStatusEntity } from '~/.server/domain/entities';
 
 export interface ClientFriendlyStatusDtoMapper {
   mapClientFriendlyStatusEntityToClientFriendlyStatusDto(clientFriendlyStatusEntity: ClientFriendlyStatusEntity): ClientFriendlyStatusDto;
