+++ frontend/app/.server/domain/mappers/federal-government-insurance-plan.dto.mapper.ts
@@ -1,7 +1,7 @@
 import { injectable } from 'inversify';
 
 import type { FederalGovernmentInsurancePlanDto } from '~/.server/domain/dtos';
-import type { FederalGovernmentInsurancePlanEntity } from '~/.server/domain/entities/federal-government-insurance-plan.entity';
+import type { FederalGovernmentInsurancePlanEntity } from '~/.server/domain/entities';
 
 export interface FederalGovernmentInsurancePlanDtoMapper {
   mapFederalGovernmentInsurancePlanEntityToFederalGovernmentInsurancePlanDto(federalGovernmentInsurancePlanEntity: FederalGovernmentInsurancePlanEntity): FederalGovernmentInsurancePlanDto;
