+++ frontend/app/.server/domain/mappers/marital-status.dto.mapper.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { MaritalStatusDto } from '~/.server/domain/dtos';
-import type { MaritalStatusEntity } from '~/.server/domain/entities/marital-status.entity';
+import type { MaritalStatusEntity } from '~/.server/domain/entities';
 
 export interface MaritalStatusDtoMapper {
   mapMaritalStatusEntityToMaritalStatusDto(maritalStatusEntity: MaritalStatusEntity): MaritalStatusDto;
