+++ frontend/app/.server/domain/mappers/preferred-communication-method.dto.mapper.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import type { ServerConfig } from '~/.server/configs/server.config';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { PreferredCommunicationMethodDto } from '~/.server/domain/dtos';
-import type { PreferredCommunicationMethodEntity } from '~/.server/domain/entities/preferred-communication-method.entity';
+import type { PreferredCommunicationMethodEntity } from '~/.server/domain/entities';
 
 export interface PreferredCommunicationMethodDtoMapper {
   mapPreferredCommunicationMethodEntityToPreferredCommunicationMethodDto(preferredCommunicationMethodEntity: PreferredCommunicationMethodEntity): PreferredCommunicationMethodDto;
