+++ frontend/app/.server/domain/mappers/provincial-government-insurance-plan.dto.mapper.ts
@@ -1,7 +1,7 @@
 import { injectable } from 'inversify';
 
 import type { ProvincialGovernmentInsurancePlanDto } from '~/.server/domain/dtos';
-import type { ProvincialGovernmentInsurancePlanEntity } from '~/.server/domain/entities/provincial-government-insurance-plan.entity';
+import type { ProvincialGovernmentInsurancePlanEntity } from '~/.server/domain/entities';
 
 export interface ProvincialGovernmentInsurancePlanDtoMapper {
   mapProvincialGovernmentInsurancePlanEntityToProvincialGovernmentInsurancePlanDto(provincialGovernmentInsurancePlanEntity: ProvincialGovernmentInsurancePlanEntity): ProvincialGovernmentInsurancePlanDto;
