+++ frontend/app/.server/domain/repositories/client-friendly-status.repository.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'inversify';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
-import type { ClientFriendlyStatusEntity } from '~/.server/domain/entities/client-friendly-status.entity';
+import type { ClientFriendlyStatusEntity } from '~/.server/domain/entities';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 import clientFriendlyStatusJsonDataSource from '~/.server/resources/power-platform/client-friendly-status.json';
