+++ frontend/app/.server/domain/repositories/country.repository.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'inversify';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
-import type { CountryEntity } from '~/.server/domain/entities/country.entity';
+import type { CountryEntity } from '~/.server/domain/entities';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 import countryJsonDataSource from '~/.server/resources/power-platform/country.json';
